requirements.txt
aniso8601==8.1.1
APScheduler==3.8.1
BareNecessities==0.2.8
bcrypt==3.2.0
blinker==1.4
bson==0.5.10
certifi==2020.12.5
cffi==1.14.5
charset-normalizer==2.0.4
click==7.1.2
cloudinary==1.25.0
dnspython==2.1.0
email-validator==1.1.2
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.10
Flask-Mail==0.9.1
flask-mongoengine==1.0.0
Flask-RESTful==0.3.8
Flask-WTF==0.14.3
gingerit==0.8.2
gunicorn==20.1.0
idna==3.1
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
mongoengine==0.22.1
pycparser==2.20
PyJWT==2.0.1
pymongo==3.11.4
python-dateutil==2.8.1
python-http-client==3.3.2
pytz==2021.1
requests==2.26.0
sendgrid==6.6.0
six==1.15.0
starkbank-ecdsa==1.1.0
urllib3==1.26.4
Werkzeug==1.0.1
WTForms==2.3.3
google-ads==12.0.0
pandas
openpyxl==3.0.7
spacy
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.0.0/en_core_web_sm-3.0.0-py3-none-any.whl